Features to close #27 and #16

+++ test/functional/jmm_get.sh
@@ -11,10 +11,19 @@ packageDir=$JMMPATH/src/github/com/jminusminus/jmmexample
 if [ -d $packageDir ]; then
 	rm -rf $packageDir
 fi
+
 jmm get github.com/jminusminus/jmmexample
 if [ ! -d $packageDir ]; then
 	echo "Package was not downloaded."
 	exit 1
 fi
+
+touch $packageDir/tmp.txt
+jmm get github.com/jminusminus/jmmexample
+if [ ! -f $packageDir/tmp.txt ]; then
+	echo "Package was downloaded when it shouldn't have been."
+	exit 1
+fi
+
 rm -rf $packageDir
 exit 0
